Refactor: div for struct for person
-
