+++ CodeStarterCamp_Week3/structForCaffeeShop.swift
@@ -7,22 +7,6 @@
 
 import Foundation
 
-class Person {
-    var name: String
-    var age: Int
-    var money: Int
-
-    init(name: String, age: Int, money: Int) {
-        self.name = name
-        self.age = age
-        self.money = money
-    }
-    
-    func order(coffee: Coffee) {
-        print("\(coffee)주세요")
-    }
-}
-
 struct CoffeeShop {
     var totalRevenue: Int
     var menu: [Coffee: Int] = [.americano: 5000, .latte: 5500, .cappuccino: 5500]
